+++ aimlflow/utils.py
@@ -81,7 +81,7 @@ def get_mlflow_experiments(client, experiment):
     return experiments
 
 
-def get_aim_run(repo_inst, run_id, experiment_name, run_cache):
+def get_aim_run(repo_inst, run_id, run_name, experiment_name, run_cache):
     if run_cache.get(run_id):
         aim_run = Run(
             run_hash=run_cache[run_id],
@@ -98,7 +98,7 @@ def get_aim_run(repo_inst, run_id, experiment_name, run_cache):
             experiment=experiment_name,
         )
         run_cache[run_id] = aim_run.hash
-
+    aim_run.name = run_name
     return aim_run
 
 
@@ -205,9 +205,8 @@ def convert_existing_logs(repo_inst, tracking_uri, experiment=None, no_cache=Fal
         runs = client.search_runs(ex.experiment_id)
 
         for run in tqdm(runs, desc=f'Parsing mlflow runs for experiment `{ex.name}`', total=len(runs)):
-            run_id = run.info.run_id
             # get corresponding `aim.Run` object for mlflow run
-            aim_run = get_aim_run(repo_inst, run_id, ex.name, run_cache)
+            aim_run = get_aim_run(repo_inst, run.info.run_id, run.info.run_name, ex.name, run_cache)
             # Collect params and tags
             collect_run_params(aim_run, run)